Alterando o retorno das classes do tipo operacao de double para bigdecimal; Alterando o nome da classe inverso para inversao; Implementando logica de calculo na tela; Adicionando visor geral para visualizar as ultimas acoes;

+++ src/br/com/clauvane/calculadora/modelo/Operacao.java
@@ -1,6 +1,8 @@
 package br.com.clauvane.calculadora.modelo;
 
+import java.math.BigDecimal;
+
 public interface Operacao {
-	double calcular(double numero);
-	double calcular(double primeiroNumero, double segundoNumero);
+	BigDecimal calcular(BigDecimal numero);
+	BigDecimal calcular(BigDecimal primeiroNumero, BigDecimal segundoNumero);
 }
